css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-targeted-element-main-frame.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title> CSS Scroll Snap Test: snap selection with targeted element</title>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/common/utils.js"></script>
</head>

<body onload="runTest()">
  <script>
    function fetchResult(key, resolve, reject) {
      const url = `/css/css-scroll-snap/snap-after-relayout/` +
        `multiple-aligned-targets/stash.py?key=${key}`;
      fetch(url).then(response => {
        return response.text();
      }).then(text => {
        if (text) {
          try {
            let result = text;
            resolve(result);
          } catch (e) {
            reject();
          }
        } else {
          fetchResult(key, resolve, reject);
        }
      });
    }

    async function runTest() {
      async function test(n) {
        return promise_test(t => new Promise(async (resolve, reject) => {
          let key = token();

          test_driver.bless("Open a URL with a text fragment directive", () => {
            window.open(`prefer-targeted-element-main-frame-target.html` +
              `?key=${key}#box${n}`, "_blank", "noopener");
          });

          fetchResult(key, resolve, reject);
        }).then(result => {
          assert_equals(result, "PASS");
        }), `targeted box${n} is selected snap target.`);
      }

      await test(1);
      await test(2);
      await test(3);
      await test(4);
      await test(5);
      await test(6);
      await test(7);
      await test(8);
      await test(9);
    }
  </script>
</body>
