LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-targeted-element-main-frame-expected.txt


FAIL targeted box is selected snap target. assert_equals: expected "9," but got "0,,Can't find variable: n,Can't find variable: n,Can't find variable: n,Can't find variable: n,Can't find variable: n,Can't find variable: n,Can't find variable: n,Can't find variable: n,Can't find variable: n"
